requirements.txt
flask
flask-cors
openai
python-dotenv
langchain-groq
langchain-core
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
